# gh-109653: Reduce import overhead of uuid module on Linux