@zooba 2024-02-08
I'm not the right person to review this.

You need someone who knows whether `sys.platform == "linux"` implies the same things as the platform module. In general, they don't, so if you want my review I would reject this change as being too risky.