@hauntsaninja 2024-02-08
Thanks for the quick response! I'll wait for someone else to review.

The value of `sys.platform` comes from https://github.com/python/cpython/blob/1032326fe46afaef57c3e01160a4f889dadfee95/configure#L4296 while the value of `platform.system` comes from `os.uname` + munging and falling back to `sys.platform`

Note we don't need a one-to-one correspondence here (and I'm not sure there is one, e.g. on ancient Python versions it was possible to have `sys.platform == "linux2"` while presumably `platform.system()` was still Linux). But that doesn't affect correctness here since we preserve the `else` branch.