@ofek 2024-02-28
This seems correct to me; I don't see any risk to this change.